TOREE-341 Fixed exception stack traces getting lost.
This is a rather ugly fix to address the issue caused by
https://issues.scala-lang.org/browse/SI-8935

The crux of the issue is that valueOfTerm returns None -- always

This introduces a small class that is bound into the REPL scope so that we
can modify its internal state and use that to retrieve values.

We can extend this to a more general case later on if needed.