Add git configuration for line ending
Git relies on user setting core.autocrlf which
is ususally unreliable as not all users are aware
of this configuration, or they forget to properly set.

The .gitattributes file is a template file with the
appropriate configuration for different file types.

After these changes are applied it's recommended to
do the following commands on your local repository:

git rm --cached -r .
git reset --hard

More info available at
https://help.github.com/articles/dealing-with-line-endings/