TOREE-378: Fix is_complete_handler responses.
This commit includes a bug fix to send indentation instead of the
response twice.

This also updates the is-complete check to require a blank line to end a
multi-line block, which is the way ipython behaves. This avoids running
code when a user intends to continue a block.

Last, this updates the indentation heuristic. First, indentation is
taken from the previous line's starting indentation (before, the first
white space was used). Then, if toree detects => or {, it adds two
spaces.