TOREE-380: Allow interpreters to produce output by MIME type.
This updates the interpreter API to return a Map of MIME type to content
instead of a String that is rendered as text. This allows interpreters
to show HTML as cell output.

This also updates the output of magic functions to be a similar MIME
type to content structure. This is cleaner and no longer requires
hacky post-processing before relaying a cell's output.