[TOREE-386] Move application name specification to run.sh
The application name of Toree's Spark driver program is currenlty
hardcoded to "Apache Toree" in the Spark context initialization
functions. By moving the name to a parameter into the launcher script,
it is possible for users to override the name by specifying an
additional "--name" parameter in the SPARK_OPTS variable (this is
because later parameters override earlier ones).