TOREE-408: Fix addJar file name test.
URI removes the query fragment from the path component and file name is
based on the path. This test used to rely on the fact that the query
fragment was included in the file name returned by URL, but this is
unreliable. The URL could easily end with some other query param instead
of the file name, so the logic wasn't reliable.