[TOREE-421] Allow thread groups (#178)
Remove the portions of Security Manager that restricted the
creation of thread groups. This change also preserves the 
prevention of calls to System.exit() from sub-threads/groups 
(including those derived from the restricted group) - otherwise 
such calls from threads in a new thread group can terminate 
the kernel.

I chose not to implement the ability to preserve the existing logic 
whereby the creation of new groups are prevented (i.e., via a new 
boolean option --prevent-new-thread-groups) since it felt like adding 
something that probably isn't necessary. If others disagree or we 
(eventually) find a good use-case for restoring that kind of functionality, 
we can add such an option at that time.

Closes #178